+++ cmd/build-all.ts
@@ -1,5 +1,5 @@
 import { join } from "node:path";
-import { detectVisualStudio2026, runLogged, copyDistributionFonts } from "./util";
+import { detectVisualStudio2026, runLogged, copyDistributionFonts, copyOpenccData } from "./util";
 import { clearDirPreserveSettings } from "./clean";
 
 let clean = false;
@@ -23,6 +23,7 @@ async function main() {
   await runLogged(msbuildPath, [sln, t, p, `/m`]);
 
   copyDistributionFonts(join("out", "rel64"));
+  copyOpenccData(join("out", "rel64"));
 
   // const outDir = join("out", "dbg64");
   // await runLogged(resolve(join(outDir, "test_util.exe")), [], outDir);